Move dynamic routers into outbound
