+++ linkerd/app/core/src/svc.rs
@@ -8,7 +8,7 @@ pub use linkerd_reconnect::NewReconnect;
 pub use linkerd_stack::{
     self as stack, layer, ArcNewService, BoxCloneService, BoxService, BoxServiceLayer, Either,
     ExtractParam, Fail, FailFast, Filter, InsertParam, MakeConnection, MapErr, MapTargetLayer,
-    NewRouter, NewService, Param, Predicate, UnwrapOr,
+    NewCloneService, NewRouter, NewService, Oneshot, Param, Predicate, UnwrapOr,
 };
 pub use linkerd_stack_tracing::{GetSpan, NewInstrument, NewInstrumentLayer};
 use stack::OnService;
