+++ linkerd/service-profiles/Cargo.toml
@@ -16,6 +16,7 @@ http = "0.2"
 http-body = "0.4"
 indexmap = "1"
 linkerd-addr = { path = "../addr" }
+linkerd-distribute = { path = "../distribute" }
 linkerd-dns-name = { path = "../dns/name" }
 linkerd-error = { path = "../error" }
 linkerd-http-box = { path = "../http-box" }
